refactor: minor changes for the review
