+++ galasa-ui/src/app/test-runs/page.tsx
@@ -10,7 +10,7 @@ import styles from "@/styles/TestRunsPage.module.css";
 import { HOME } from "@/utils/constants/breadcrumb";
 import { Suspense } from "react";
 import { getYesterday } from "@/utils/timeOperations";
-import { fetchAllTestRunsByPaging, getRequestorList, getResultsNames } from '@/lib/data/testRuns';
+import { fetchAllTestRunsByPaging, getRequestorList, getResultsNames } from '@/utils/testRuns';
 
 
 export default async function TestRunsPage({searchParams}: {searchParams: {[key: string]: string | undefined}} ) {
