+++ galasa-ui/src/tests/functions/fetchAllTestRunsByPaging.test.tsx
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-import { fetchAllTestRunsByPaging } from '@/lib/data/testRuns';
+import { fetchAllTestRunsByPaging } from '@/utils/testRuns';
 
 import { ResultArchiveStoreAPIApi, Run } from '@/generated/galasaapi';
 import { createAuthenticatedApiConfiguration } from '@/utils/api';
